external workload: move GCV2 abort helper
